requirements.txt
# Your existing dependencies
# (keep whatever you already have)

# Document Similarity Dependencies
sentence-transformers==2.2.2
torch>=1.9.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
tqdm==4.65.0

# Streamlit for frontend
streamlit==1.28.1
plotly==5.17.0

# File handling
Pillow==10.0.0

# Optional: For better performance
transformers>=4.21.0